Remove requirement for jenkins credentials
The jenkins credentials are not required to query the jobs
and are somewhat painful to maintain. Rather than use them
to derive an email address, we add the email_from parameter
and remove the use of jenkins credentials entirely.

The email_from argument is only required if an email is
being sent.

Fixes: #178